Add files via upload
Research Objectives 

This study aims to investigate the levels of acculturation, stigma tolerance, self-efficacy, expertness, and the mental health impact of COVID-19 on Saudi university students in the United States. It also seeks to understand the correlations between these factors and their collective role in predicting the intention to seek mental health help. By examining these relationships, the research will shed light on the key predictors of mental health help-seeking intentions, both in isolation and when accounting for the impact of COVID-19 and symptoms of anxiety and depression. Each of the following questions will be checking by survey conducted on (n=196) participants.

i-	What are the levels of acculturation, stigma tolerance, expertness, self-efficacy, intent, COVID-19 impact and experience, anxiety, and depression among Saudi University students in the United States? 

ii-	What are the bivariate correlations between acculturation, stigma tolerance, expertness, self-efficacy, and mental health help-seeking intention?

iii-	Are acculturation, stigma tolerance, expertness, and self-efficacy significant predictors of mental health help-seeking intentions?

iv-	Are acculturation, stigma tolerance and expertness, and self-efficacy significant predictors of mental health help-seeking intention after adjusting for the impact and experience of COVID-19 and current level of anxiety and depression?